remove timer, move progress dialog to mainwindow
- forward FileTransfer signals from SessionManager to MainWindow
- removed majority of Qt::DirectConnection connect calls, but one :
at sessionmanager.cpp:217

    connect(this, &SessionManager::progressDialogCancelled,
            file_transfer, &FileTransfer::cancelTransfer,
            Qt::DirectConnection);
The fact that FileTransfer uses QThread is an implementation detail of
FileTransfer. The thread is actually started when and only when the file
transfer protocol begins. But at the time and space of the signal/slot
connection, we don't know if the transfer will really take place (chosen
file could be empty, or not readable for example).